% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordsToBearing.R
\name{angleDiff}
\alias{angleDiff}
\title{Calculate the signed difference between angles (in degrees)}
\usage{
angleDiff(angle1, angle2)
}
\arguments{
\item{angle1}{Source angle}

\item{angle2}{Target angle}
}
\value{
The signed difference between the two angles
}
\description{
Calculate the signed difference between angles (in degrees)
}

