% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrajectory.R
\name{as.data.frame.flighttrajectory}
\alias{as.data.frame.flighttrajectory}
\title{Convert a trajectory to a data.frame}
\usage{
\method{as.data.frame}{flighttrajectory}(x)
}
\description{
Convert a trajectory to a data.frame
}

