% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercePath.R
\name{as.flightpath}
\alias{as.flightpath}
\title{Attempt to coerce an object into an instance of \code{flightpath}.}
\usage{
as.flightpath(x)
}
\description{
Attempt to coerce an object into an instance of \code{flightpath}.
}

