% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceTrajectory.R
\name{as.flighttrajectory}
\alias{as.flighttrajectory}
\title{Attempt to coerce an object into an instance of \code{flighttrajectory}.}
\usage{
as.flighttrajectory(x)
}
\description{
Attempt to coerce an object into an instance of \code{flighttrajectory}.
}

