% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordsToBearing.R
\name{coordsToBearing}
\alias{coordsToBearing}
\title{Calculate the bearing between successive longitude/latitude points.}
\usage{
coordsToBearing(longitude, latitude)
}
\arguments{
\item{longitude}{Latitude values.}

\item{latitude}{Longitude values.}
}
\value{
The numeric vecotor giving the instantaneious bearing at each point
  along the trajectory, in degrees. This will have the same length as the
  trajectory, but the last element will be \code{NA}.
}
\description{
Calculate the bearing between successive longitude/latitude points.
}

