% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPath.R
\name{createPath}
\alias{createPath}
\title{Create a flightpath object from coordinate info.}
\usage{
createPath(longitude, latitude, altitude = NA)
}
\arguments{
\item{longitude}{Required; numeric vector giving aircraft longitude in
degrees.}

\item{latitude}{Required; numeric vector giving aircraft latitude in degrees.}

\item{altitude}{Optional; numeric vector giving aircraft altitude (AGL) in
feet.}
}
\value{
A flightpath object encapsulating these parameters (with default
  values substituded as necessary).
}
\description{
Create a flightpath object from coordinate info.
}
\details{
\code{longitude} and \code{latitude} must be the same length.
  \code{altitude} must also have a length equal to these parameters or be
  scalar. If \code{altitude} is \code{NA}, operations that compare values to
  flightpaths will be affected.
}

