% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrajectory.R
\name{createTrajectory}
\alias{createTrajectory}
\title{Create a flighttrajectory object from the flight info.}
\usage{
createTrajectory(longitude, latitude, altitude = 0, timestamp = NULL,
  bearing = NULL, groundspeed = NULL)
}
\arguments{
\item{longitude}{Required; numeric vector giving aircraft longitude in
degrees.}

\item{latitude}{Required; numeric vector giving aircraft latitude in degrees.}

\item{altitude}{Optional; numeric vector giving aircraft altitude (AGL) in
feet. If missing, it will be set to 0.}

\item{timestamp}{Optional; numeric vector giving the time of each observation
in seconds. If missing, the observation period is assumed to be 1 s.}

\item{bearing}{Optional; numeric vector giving the current bearing in
degrees. If missing, it is estimated using pairs of successive lon/lat
observations.}

\item{groundspeed}{Optional; numeric vector giving the current ground speed
of the aircraft in knots. If missing, it is estimated using pairs of
successive lon/lat observations.}
}
\value{
A flighttrajectory object encapsulating these parameters (with
  default values substituded as necessary).
}
\description{
Create a flighttrajectory object from the flight info.
}
\details{
\code{longitude} and \code{latitude} must be the same length.
  \code{timestamp}, \code{bearing}, and \code{groundspeed}, if present, must
  also match this length. \code{altitude} must also have a length equal to
  these parameters or be scalar.
}

