% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFromPath.R
\name{distanceFromPath}
\alias{distanceFromPath}
\title{Calculate the distance of a flight trajectory from a flight path.}
\usage{
distanceFromPath(trajectory, path)
}
\arguments{
\item{trajectory}{A matrix or SpatialPoints object indicating the trajectory
of an aircraft.}

\item{path}{A matrix or SpatialPoints object indicating the ordered waypoints
a pre-defined flight path.}
}
\value{
A data.frame containing two columns representing the distance between
  the aircraft and its planned flight path (in feet): \code{horizontal}
  indicates the horizontal distance and \code{vertical} indicates the
  vertical distance.
}
\description{
Calculate the distance of a flight trajectory from a flight path.
}

