% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpathr.R
\docType{package}
\name{flightpathr}
\alias{flightpathr}
\alias{flightpathr-package}
\title{flightpathr: Tools to analyze aircraft and flight path data.}
\description{
This package provides functions for analyzing flight track and flight path
data. Useful for studies of pilot and controller performance.
}

