% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3dCoords.R
\name{get3dCoords}
\alias{get3dCoords}
\title{Checks for appropriate input and coerces it into a nx3 coordinate matrix.}
\usage{
get3dCoords(obj)
}
\arguments{
\item{obj}{An object that hopefully contains some coordinates.}
}
\value{
A nx3 coordinate matrix
}
\description{
Checks for appropriate input and coerces it into a nx3 coordinate matrix.
}

