% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{hysteresisThresh}
\alias{hysteresisThresh}
\title{Helper function for hysteresis thresholding: if a group of neighboring
segments are all above the low threshold and neighbor a segment above the
high threshold, they're "above threshold".}
\usage{
hysteresisThresh(x, lo, hi = NA)
}
\description{
Helper function for hysteresis thresholding: if a group of neighboring
segments are all above the low threshold and neighbor a segment above the
high threshold, they're "above threshold".
}

