% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{identifyBearingChanges}
\alias{identifyBearingChanges}
\title{Identify the timepoints in a trajectory that correspond to a change in
commanded heading.}
\usage{
identifyBearingChanges(trajectory, hiThresh)
}
\arguments{
\item{trajectory}{A \code{flighttrajectory} object (or input coercable to
one) indicating the trajectory of an aircraft.}

\item{hiThresh}{A bearing change (in degrees); any time point associated with
a change in bearing greater than this value will definitely be labeled a
turn.}
}
\value{
A logical vector indicating whether each timepoint can be considered
  a turn.
}
\description{
Identify the timepoints in a trajectory that correspond to a change in
commanded heading.
}

