% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{identifyChanges}
\alias{identifyChanges}
\title{Helper function for identifying changes based on change magnitude and change
slope.}
\usage{
identifyChanges(x, dx, xThresh, dxThresh)
}
\description{
Helper function for identifying changes based on change magnitude and change
slope.
}

