% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{identifyHeadingChanges}
\alias{identifyHeadingChanges}
\title{Identify the timepoints in a trajectory that correspond to a change in
commanded heading.}
\usage{
identifyHeadingChanges(trajectory, hiThresh)
}
\arguments{
\item{trajectory}{A matrix or SpatialPoints object indicating the trajectory
of an aircraft.}

\item{hiThresh}{A heading change (in degrees); any time point associated with
a change in heading greater than this value will definitely be labeled a
turn.}
}
\value{
A logical vector indicating whether each timepoint can be considered
  a turn.
}
\description{
Identify the timepoints in a trajectory that correspond to a change in
commanded heading.
}

