% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{identifySpeedChanges}
\alias{identifySpeedChanges}
\title{Identify the timepoints in a trajectory that correspond to a change in
speed.}
\usage{
identifySpeedChanges(trajectory, speedThresh, accelerationThresh)
}
\arguments{
\item{trajectory}{A \code{flighttrajectory} object (or input coercable to
one) indicating the trajectory of an aircraft.}

\item{speedThresh}{A speed threshold (in knots).}

\item{accelerationThresh}{An accelaration threshold (in knots per second).}
}
\value{
A logical vector indicating whether each timepoint can be considered
  a change in speed
}
\description{
Identify the timepoints in a trajectory that correspond to a change in
speed.
}

