% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateTrajectory.R
\name{interpolateTrajectory}
\alias{interpolateTrajectory}
\title{Interpolate a trajectory (in time)}
\usage{
interpolateTrajectory(trajectory, timestamp)
}
\arguments{
\item{trajectory}{A \code{flighttrajectory} object.}

\item{timestamp}{The new timestamp along which the data should be 
interpolated.}
}
\value{
A new \code{flighttrajectory} with the given \code{timestamp}
}
\description{
Interpolate a trajectory (in time)
}
\details{
This just performs linear interpolation for all of the values in the
  trajectory. A better approach would make use of the bearing and velocity
  information to smoothly interpolate the coordinates.
}

