% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3dCoords.R
\name{get3dCoords}
\alias{get3dCoords}
\title{Checks for a \code{flighttrajectory} or \code{flightpath} object and returns
it's position in a 3D matrix.}
\usage{
get3dCoords(coordObject)
}
\arguments{
\item{coordObject}{An object that hopefully contains some coordinates.}
}
\value{
A nx3 coordinate matrix giving longitude/latitude/altitude.
}
\description{
Checks for a \code{flighttrajectory} or \code{flightpath} object and returns
it's position in a 3D matrix.
}
\details{
It'll be safer to access \code{flighttrajectory} and
  \code{flightpath} objects' coordinates through this.
}

