% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{identifyAltitudeChanges}
\alias{identifyAltitudeChanges}
\title{Identify the timepoints in a trajectory that correspond to a change in
altitude.}
\usage{
identifyAltitudeChanges(trajectory, altitudeThresh, verticalSpeedThresh)
}
\arguments{
\item{trajectory}{A \code{flighttrajectory} object (or input coercable to
one) indicating the trajectory of an aircraft.}

\item{altitudeThresh}{An altitude threshold (in feet).}

\item{verticalSpeedThresh}{A vertical speed threshold (in feet per second).}
}
\value{
A logical vector indicating whether each timepoint can be considered
  a climb or descent.
}
\description{
Identify the timepoints in a trajectory that correspond to a change in
altitude.
}

