% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{identifyBearingChanges}
\alias{identifyBearingChanges}
\title{Identify the timepoints in a trajectory that correspond to a change in
commanded heading.}
\usage{
identifyBearingChanges(trajectory, bearingThresh, turnThresh)
}
\arguments{
\item{trajectory}{A \code{flighttrajectory} object (or input coercable to
one) indicating the trajectory of an aircraft.}

\item{bearingThresh}{A bearing change threshold (in degrees).}

\item{turnThresh}{A turn rate threshold (in degrees per second).}
}
\value{
A logical vector indicating whether each timepoint can be considered
  a turn.
}
\description{
Identify the timepoints in a trajectory that correspond to a change in
commanded heading.
}
\details{
A period is marked as a turn if the turn rate is greater than
  \code{turnThresh} and the total bearing change is greater than
  \code{bearingThresh}.
}

