% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyManeuvers.R
\name{identifySpeedChanges}
\alias{identifySpeedChanges}
\title{Identify the timepoints in a trajectory that correspond to a change in
altitude.}
\usage{
identifySpeedChanges(trajectory, loThresh, hiThresh = NA)
}
\arguments{
\item{trajectory}{A \code{flighttrajectory} object (or input coercable to
one) indicating the trajectory of an aircraft.}

\item{hiThresh}{An altitude change (in feet per second); any time point
associated with a change in altitude greater than this value will
definitely be labeled a climb or descent.}
}
\value{
A logical vector indicating whether each timepoint can be considered
  a climb or descent.
}
\description{
Identify the timepoints in a trajectory that correspond to a change in
altitude.
}

